Create R script file in workspace
Trying to use the r step inside a docker container step (like
withDockerContainer) in a pipeline script failed because the R script
file was created inside the /tmp folder of the jenkins node instead of
the docker container. The /tmp folder is not bound to the docker
container as a volume by default, but the project workspace is.